make iterableEquality private
