+++ expect/_matchers.ts
@@ -17,8 +17,7 @@ import { format } from "./_format.ts";
 import type { AnyConstructor, MatcherContext, MatchResult } from "./_types.ts";
 import { getMockCalls } from "./_mock_util.ts";
 import { inspectArg, inspectArgs } from "./_inspect_args.ts";
-import { iterableEquality } from "./iterable_equality.ts";
-import { buildEqualOptions } from "./_utils.ts";
+import { buildEqualOptions, iterableEquality } from "./_utils.ts";
 
 export function toBe(context: MatcherContext, expect: unknown): MatchResult {
   if (context.isNot) {
